backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
pydantic[email]>=2.10.0
python-jose[cryptography]>=3.3.0
asyncpg>=0.30.0
supabase>=2.10.0
resend>=2.4.0
python-multipart>=0.0.20
httpx>=0.28.0
jinja2>=3.1.5
slowapi>=0.1.9
passlib[bcrypt]>=1.7.4
zxcvbn>=4.4.28

# Dev
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
python-dotenv>=1.0.0
ruff>=0.8.0
